move migrators
