+++ dbot.py
@@ -247,6 +247,6 @@ for cog in STATIC_COGS:
 
 if __name__ == '__main__':
     bot.state = "run"
-    if not bot.rdb.exists('build_num'): bot.rdb.set('build_num', 114)  # this was added in build 114
+    if not bot.rdb.exists('build_num'): bot.rdb.set('build_num', 0)
     bot.rdb.incr('build_num')
     bot.run(bot.credentials.token)
